+++ modules/sr810.c
@@ -435,7 +435,7 @@ lockin_get_data( Var_T * v )
         if ( num_channels == 1 )
             return vars_push( FLOAT_VAR, data[ 0 ] );
         else
-            return vars_push( FLOAT_ARR, data, num_channels );
+            return vars_push( FLOAT_ARR, data, ( long ) num_channels );
     }
 
     /* If we need less than two channels we've got to pass the function an
@@ -452,7 +452,7 @@ lockin_get_data( Var_T * v )
     if ( using_dummy_channels )
         return vars_push( FLOAT_VAR, data[ 0 ] );
 
-    return vars_push( FLOAT_ARR, data, num_channels );
+    return vars_push( FLOAT_ARR, data, ( long ) num_channels );
 }
