+++ src/vars_add.c
@@ -169,7 +169,7 @@ vars_int_var_add( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( INT_ARR, v2->val.lpnt, v2->len );
+                new_var = vars_push( INT_ARR, v2->val.lpnt, ( long ) v2->len );
 
             if ( v1->val.lval != 0 )
                 for ( i = 0; i < v2->len; i++ )
@@ -185,7 +185,8 @@ vars_int_var_add( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( FLOAT_ARR, v2->val.dpnt, v2->len );
+                new_var = vars_push( FLOAT_ARR, v2->val.dpnt,
+                                     ( long ) v2->len );
 
             if ( v1->val.lval != 0 )
                 for ( i = 0; i < v2->len; i++ )
@@ -265,7 +266,7 @@ vars_float_var_add( Var_T * v1,
             break;
 
         case INT_ARR :
-            new_var = vars_push( FLOAT_ARR, NULL, v2->len );
+            new_var = vars_push( FLOAT_ARR, NULL, ( long ) v2->len );
 
             for ( i = 0; i < new_var->len; i++ )
                 new_var->val.dpnt[ i ]
@@ -280,7 +281,8 @@ vars_float_var_add( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( FLOAT_ARR, v2->val.dpnt, v2->len );
+                new_var = vars_push( FLOAT_ARR, v2->val.dpnt,
+                                     ( long ) v2->len );
 
             if ( v1->val.dval != 0.0 )
                 for ( i = 0; i < new_var->len; i++ )
@@ -363,7 +365,7 @@ vars_int_arr_add( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( INT_ARR, v2->val.lpnt, v2->len );
+                new_var = vars_push( INT_ARR, v2->val.lpnt, ( long ) v2->len );
 
             for ( i = 0; i < new_var->len; i++ )
                 new_var->val.lpnt[ i ] += v1->val.lpnt[ i ];
@@ -378,7 +380,8 @@ vars_int_arr_add( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( FLOAT_ARR, v2->val.dpnt, v2->len );
+                new_var = vars_push( FLOAT_ARR, v2->val.dpnt,
+                                     ( long ) v2->len );
 
             for ( i = 0; i < v1->len; i++ )
                 new_var->val.dpnt[ i ] += ( double ) v1->val.lpnt[ i ];
@@ -397,7 +400,7 @@ vars_int_arr_add( Var_T * v1,
                 new_var = vars_push( v2->type, v2 );
 
             for ( i = 0; i < new_var->len; i++ )
-                vars_add( vars_push( INT_ARR, v1->val.lpnt, v1->len ),
+                vars_add( vars_push( INT_ARR, v1->val.lpnt, ( long ) v1->len ),
                           new_var->val.vptr[ i ] );
 
             vars_pop( v1 );
@@ -458,7 +461,8 @@ vars_float_arr_add( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( FLOAT_ARR, v2->val.dpnt, v2->len );
+                new_var = vars_push( FLOAT_ARR, v2->val.dpnt,
+                                     ( long ) v2->len );
 
             for ( i = 0; i < new_var->len; i++ )
                 new_var->val.dpnt[ i ] += v1->val.dpnt[ i ];
@@ -477,7 +481,8 @@ vars_float_arr_add( Var_T * v1,
                 new_var = vars_push( v2->type, v2 );
 
             for ( i = 0; i < new_var->len; i++ )
-                vars_add( vars_push( FLOAT_ARR, v1->val.dpnt, v1->len ),
+                vars_add( vars_push( FLOAT_ARR, v1->val.dpnt,
+                                     ( long ) v1->len ),
                           new_var->val.vptr[ i ] );
 
             vars_pop( v1 );
