+++ src/vars_mod.c
@@ -177,7 +177,7 @@ vars_int_var_mod( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( INT_ARR, v2->val.lpnt, v2->len );
+                new_var = vars_push( INT_ARR, v2->val.lpnt, ( long ) v2->len );
 
             for ( i = 0; i < v2->len; i++ )
             {
@@ -204,7 +204,8 @@ vars_int_var_mod( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( FLOAT_ARR, v2->val.dpnt, v2->len );
+                new_var = vars_push( FLOAT_ARR, v2->val.dpnt,
+                                     ( long ) v2->len );
 
             for ( i = 0; i < v2->len; i++ )
             {
@@ -329,7 +330,7 @@ vars_float_var_mod( Var_T * v1,
             break;
 
         case INT_ARR :
-            new_var = vars_push( FLOAT_ARR, NULL, v2->len );
+            new_var = vars_push( FLOAT_ARR, NULL, ( long ) v2->len );
 
             for ( i = 0; i < new_var->len; i++ )
             {
@@ -355,7 +356,8 @@ vars_float_var_mod( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( FLOAT_ARR, v2->val.dpnt, v2->len );
+                new_var = vars_push( FLOAT_ARR, v2->val.dpnt,
+                                     ( long ) v2->len );
 
             for ( i = 0; i < new_var->len; i++ )
             {
@@ -474,7 +476,7 @@ vars_int_arr_mod( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( INT_ARR, v2->val.lpnt, v2->len );
+                new_var = vars_push( INT_ARR, v2->val.lpnt, ( long ) v2->len );
 
             for ( i = 0; i < new_var->len; i++ )
             {
@@ -501,7 +503,8 @@ vars_int_arr_mod( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( FLOAT_ARR, v2->val.dpnt, v2->len );
+                new_var = vars_push( FLOAT_ARR, v2->val.dpnt,
+                                     ( long ) v2->len );
 
             for ( i = 0; i < v1->len; i++ )
             {
@@ -534,7 +537,8 @@ vars_int_arr_mod( Var_T * v1,
                 new_var = vars_push( v2->type, v2 );
 
             for ( i = 0; i < new_var->len; i++ )
-                vars_mod_i( vars_push( INT_ARR, v1->val.lpnt, v1->len ),
+                vars_mod_i( vars_push( INT_ARR, v1->val.lpnt,
+                                       ( long ) v1->len ),
                             new_var->val.vptr[ i ], exc );
 
             vars_pop( v1 );
@@ -592,7 +596,8 @@ vars_float_arr_mod( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( FLOAT_ARR, v2->val.dpnt, v2->len );
+                new_var = vars_push( FLOAT_ARR, v2->val.dpnt,
+                                     ( long ) v2->len );
 
             for ( i = 0; i < new_var->len; i++ )
             {
@@ -623,7 +628,8 @@ vars_float_arr_mod( Var_T * v1,
                 new_var = vars_push( v2->type, v2 );
 
             for ( i = 0; i < new_var->len; i++ )
-                vars_mod_i( vars_push( FLOAT_ARR, v1->val.dpnt, v1->len ),
+                vars_mod_i( vars_push( FLOAT_ARR, v1->val.dpnt,
+                                       ( long ) v1->len ),
                             new_var->val.vptr[ i ], exc );
 
             vars_pop( v1 );
