+++ src/vars_mult.c
@@ -132,7 +132,7 @@ vars_int_var_mult( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( INT_ARR, v2->val.lpnt, v2->len );
+                new_var = vars_push( INT_ARR, v2->val.lpnt, ( long ) v2->len );
 
             if ( v1->val.lval != 1 )
                 for ( i = 0; i < v2->len; i++ )
@@ -148,7 +148,8 @@ vars_int_var_mult( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( FLOAT_ARR, v2->val.dpnt, v2->len );
+                new_var = vars_push( FLOAT_ARR, v2->val.dpnt,
+                                     ( long ) v2->len );
 
             if ( v1->val.lval != 1 )
                 for ( i = 0; i < v2->len; i++ )
@@ -226,7 +227,7 @@ vars_float_var_mult( Var_T * v1,
             break;
 
         case INT_ARR :
-            new_var = vars_push( FLOAT_ARR, NULL, v2->len );
+            new_var = vars_push( FLOAT_ARR, NULL, ( long ) v2->len );
 
             for ( i = 0; i < new_var->len; i++ )
                 new_var->val.dpnt[ i ]
@@ -241,7 +242,8 @@ vars_float_var_mult( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( FLOAT_ARR, v2->val.dpnt, v2->len );
+                new_var = vars_push( FLOAT_ARR, v2->val.dpnt,
+                                     ( long ) v2->len );
 
             if ( v1->val.dval != 1.0 )
                 for ( i = 0; i < new_var->len; i++ )
@@ -323,7 +325,7 @@ vars_int_arr_mult( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( INT_ARR, v2->val.lpnt, v2->len );
+                new_var = vars_push( INT_ARR, v2->val.lpnt, ( long ) v2->len );
 
             for ( i = 0; i < new_var->len; i++ )
                 new_var->val.lpnt[ i ] *= v1->val.lpnt[ i ];
@@ -338,7 +340,8 @@ vars_int_arr_mult( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( FLOAT_ARR, v2->val.dpnt, v2->len );
+                new_var = vars_push( FLOAT_ARR, v2->val.dpnt,
+                                     ( long ) v2->len );
 
             for ( i = 0; i < v1->len; i++ )
                 new_var->val.dpnt[ i ] *= ( double ) v1->val.lpnt[ i ];
@@ -357,7 +360,7 @@ vars_int_arr_mult( Var_T * v1,
                 new_var = vars_push( v2->type, v2 );
 
             for ( i = 0; i < new_var->len; i++ )
-                vars_mult( vars_push( INT_ARR, v1->val.lpnt, v1->len ),
+                vars_mult( vars_push( INT_ARR, v1->val.lpnt, ( long ) v1->len ),
                            new_var->val.vptr[ i ] );
 
             vars_pop( v1 );
@@ -412,7 +415,8 @@ vars_float_arr_mult( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( FLOAT_ARR, v2->val.dpnt, v2->len );
+                new_var = vars_push( FLOAT_ARR, v2->val.dpnt,
+                                     ( long ) v2->len );
 
             for ( i = 0; i < new_var->len; i++ )
                 new_var->val.dpnt[ i ] *= v1->val.dpnt[ i ];
@@ -431,7 +435,8 @@ vars_float_arr_mult( Var_T * v1,
                 new_var = vars_push( v2->type, v2 );
 
             for ( i = 0; i < new_var->len; i++ )
-                vars_mult( vars_push( FLOAT_ARR, v1->val.dpnt, v1->len ),
+                vars_mult( vars_push( FLOAT_ARR, v1->val.dpnt,
+                                      ( long ) v1->len ),
                            new_var->val.vptr[ i ] );
 
             vars_pop( v1 );
