+++ src/vars_pow.c
@@ -174,7 +174,7 @@ vars_int_var_pow( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( INT_ARR, v2->val.lpnt, v2->len );
+                new_var = vars_push( INT_ARR, v2->val.lpnt, ( long ) v2->len );
 
             for ( i = 0; i < v2->len; i++ )
             {
@@ -197,7 +197,8 @@ vars_int_var_pow( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( FLOAT_ARR, v2->val.dpnt, v2->len );
+                new_var = vars_push( FLOAT_ARR, v2->val.dpnt,
+                                     ( long ) v2->len );
 
             for ( i = 0; i < v2->len; i++ )
             {
@@ -322,7 +323,7 @@ vars_float_var_pow( Var_T * v1,
             break;
 
         case INT_ARR :
-            new_var = vars_push( FLOAT_ARR, NULL, v2->len );
+            new_var = vars_push( FLOAT_ARR, NULL, ( long ) v2->len );
 
             for ( i = 0; i < new_var->len; i++ )
             {
@@ -350,7 +351,8 @@ vars_float_var_pow( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( FLOAT_ARR, v2->val.dpnt, v2->len );
+                new_var = vars_push( FLOAT_ARR, v2->val.dpnt,
+                                     ( long ) v2->len );
 
             for ( i = 0; i < new_var->len; i++ )
             {
@@ -469,7 +471,7 @@ vars_int_arr_pow( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( INT_ARR, v2->val.lpnt, v2->len );
+                new_var = vars_push( INT_ARR, v2->val.lpnt, ( long ) v2->len );
 
             for ( i = 0; i < new_var->len; i++ )
             {
@@ -492,7 +494,8 @@ vars_int_arr_pow( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( FLOAT_ARR, v2->val.dpnt, v2->len );
+                new_var = vars_push( FLOAT_ARR, v2->val.dpnt,
+                                     ( long ) v2->len );
 
             for ( i = 0; i < v1->len; i++ )
             {
@@ -527,7 +530,8 @@ vars_int_arr_pow( Var_T * v1,
                 new_var = vars_push( v2->type, v2 );
 
             for ( i = 0; i < new_var->len; i++ )
-                vars_pow_i( vars_push( INT_ARR, v1->val.lpnt, v1->len ),
+                vars_pow_i( vars_push( INT_ARR, v1->val.lpnt,
+                                       ( long ) v1->len ),
                             new_var->val.vptr[ i ], exc );
 
             vars_pop( v1 );
@@ -585,7 +589,8 @@ vars_float_arr_pow( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( FLOAT_ARR, v2->val.dpnt, v2->len );
+                new_var = vars_push( FLOAT_ARR, v2->val.dpnt,
+                                     ( long ) v2->len );
 
             for ( i = 0; i < new_var->len; i++ )
             {
@@ -618,7 +623,8 @@ vars_float_arr_pow( Var_T * v1,
                 new_var = vars_push( v2->type, v2 );
 
             for ( i = 0; i < new_var->len; i++ )
-                vars_pow_i( vars_push( FLOAT_ARR, v1->val.dpnt, v1->len ),
+                vars_pow_i( vars_push( FLOAT_ARR, v1->val.dpnt,
+                                       ( long ) v1->len ),
                             new_var->val.vptr[ i ], exc );
 
             vars_pop( v1 );
