+++ src/vars_sub.c
@@ -164,7 +164,7 @@ vars_int_var_sub( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( INT_ARR, v2->val.lpnt, v2->len );
+                new_var = vars_push( INT_ARR, v2->val.lpnt, ( long ) v2->len );
 
             for ( i = 0; i < v2->len; i++ )
             {
@@ -185,7 +185,8 @@ vars_int_var_sub( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( FLOAT_ARR, v2->val.dpnt, v2->len );
+                new_var = vars_push( FLOAT_ARR, v2->val.dpnt,
+                                     ( long ) v2->len );
 
             for ( i = 0; i < v2->len; i++ )
             {
@@ -284,7 +285,7 @@ vars_float_var_sub( Var_T * v1,
             break;
 
         case INT_ARR :
-            new_var = vars_push( FLOAT_ARR, NULL, v2->len );
+            new_var = vars_push( FLOAT_ARR, NULL, ( long ) v2->len );
 
             for ( i = 0; i < new_var->len; i++ )
             {
@@ -304,7 +305,8 @@ vars_float_var_sub( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( FLOAT_ARR, v2->val.dpnt, v2->len );
+                new_var = vars_push( FLOAT_ARR, v2->val.dpnt,
+                                     ( long ) v2->len );
 
             for ( i = 0; i < new_var->len; i++ )
             {
@@ -405,7 +407,7 @@ vars_int_arr_sub( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( INT_ARR, v2->val.lpnt, v2->len );
+                new_var = vars_push( INT_ARR, v2->val.lpnt, ( long ) v2->len );
 
             for ( i = 0; i < new_var->len; i++ )
             {
@@ -426,7 +428,8 @@ vars_int_arr_sub( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( FLOAT_ARR, v2->val.dpnt, v2->len );
+                new_var = vars_push( FLOAT_ARR, v2->val.dpnt,
+                                     ( long ) v2->len );
 
             for ( i = 0; i < v1->len; i++ )
             {
@@ -451,7 +454,8 @@ vars_int_arr_sub( Var_T * v1,
                 new_var = vars_push( v2->type, v2 );
 
             for ( i = 0; i < new_var->len; i++ )
-                vars_sub_i( vars_push( INT_ARR, v1->val.lpnt, v1->len ),
+                vars_sub_i( vars_push( INT_ARR, v1->val.lpnt,
+                                       ( long ) v1->len ),
                             new_var->val.vptr[ i ], exc );
 
             vars_pop( v1 );
@@ -509,7 +513,8 @@ vars_float_arr_sub( Var_T * v1,
             if ( v2->flags & IS_TEMP )
                 new_var = v2;
             else
-                new_var = vars_push( FLOAT_ARR, v2->val.dpnt, v2->len );
+                new_var = vars_push( FLOAT_ARR, v2->val.dpnt,
+                                     ( long ) v2->len );
 
             for ( i = 0; i < new_var->len; i++ )
             {
@@ -534,7 +539,8 @@ vars_float_arr_sub( Var_T * v1,
                 new_var = vars_push( v2->type, v2 );
 
             for ( i = 0; i < new_var->len; i++ )
-                vars_sub_i( vars_push( FLOAT_ARR, v1->val.dpnt, v1->len ),
+                vars_sub_i( vars_push( FLOAT_ARR, v1->val.dpnt,
+                                       ( long ) v1->len ),
                             new_var->val.vptr[ i ], exc );
 
             vars_pop( v1 );
